requirements.txt
openai
streamlit
python-dotenv
pypdf2